journald input is only compiled on Linux, move test helpers
Journald input only works on Linux because Systemd is only available
for Linux, so remove any file from the journald input that are built
on OSes other than Linux.

All integration test helpers from system-logs input are moved to a
file that is build on all platforms.
